% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{trainCR7}
\alias{trainCR7}
\title{Training Football League Matches Across the World}
\usage{
trainCR7(dataset = leagueCR7, condition)
}
\arguments{
\item{dataset}{Football leagues dataset: default of leagueCR7::leagueCR7}

\item{condition}{The name of the league}
}
\description{
The package made it available multinomial regresion models that helps you to train footbal league matches from  leagueCR7 dataset. The dataset can be install in R using remotes::install_github("Franosei/Football_dataset_package/leagueCR7")
}
